transfer

+++ src/app/Controllers/WorkController.php
@@ -9,13 +9,15 @@ use Bank\DB\JsonDb;
 class WorkController
 {
     public static function allAccounts(){
-        return App::view('work', ['title' => 'Work', 'accounts'=> DataController::getDB()]);
+
+        return App::view('work', ['title' => 'Work', 'accounts'=> DataController::getDB(), 'messages' => MessagesController::get()]);
     }
     public static function allAccountsApi(): void
     {
         echo json_encode(DataController::getDB()->showAll());
     }
     public static function deleteUserApi($userId){
+
             DataController::deleteUserById($userId);
     }
     public static function editApi($id){
@@ -37,20 +39,49 @@ class WorkController
         }
          App::redirect();
     }
+
     public static function transfer($userId){
         if (AuthorityController::auth()){
-        return App::view('transfer', ['title' => 'transfer', 'user'=> DataController::getUserById($userId)]);
+        return App::view('transfer', ['title' => 'transfer', 'user'=> DataController::getUserById($userId), 'messages'=> MessagesController::get()]);
         }
          App::redirect();
     }
+    public static function transferComplete(){
+
+    foreach (DataController::getDB()->showAll() as $value){
+        if ($value->credit_card == $_POST['cc']){
+            $toUser = $value;
+        }
+    }
+        $user = DataController::getUserById($_POST['i']);
+        $toUser = (array)$toUser;
+        $moneyTxt = (substr($user['money'] ?? '€123',3));
+        $money = (int)str_replace(',','',$moneyTxt) / 100;
+        $moneyToUser = (substr($toUser['money'] ?? '€123',3));
+        $moneyTo = (int)str_replace(',','',$moneyToUser) / 100;
+        $money -= $_POST['sum'];
+        $moneyTo += $_POST['sum'];
+        $user['money'] = '€'."$money";
+        $toUser['money'] = '€'."$moneyTo";
+        DataController::saveUser($user['id'], $user);
+        DataController::saveUser($toUser['id'], $toUser);
+        MessagesController::add('All success', 'success');
+        return (new HomeController())->transfer();
+    }
+
     public static function edit(){
         return App::view('editedView', ['title' => 'Edited']);
     }
     public static function deleteUser($userId){
-        if (AuthorityController::auth()) {
+        $user = DataController::getUserById($userId);
+        if (AuthorityController::auth() && Conditions::haveMoney($userId)) {
             DataController::deleteUserById($userId);
              App::redirect('work');
+        } else {
+            MessagesController::add($user['first_name'].' '.'still have money', 'alert');
+            App::redirect('work');
         }
+
     }
     public static function createOpen(){
         return App::view('create', ['title' => 'create']);
